app/requirements.txt
git+https://github.com/openai/CLIP.git
fastapi==0.116.1
numpy==2.3.1
pandas==2.3.1
Pillow==11.3.0
pydantic==2.11.7
python-dotenv==1.1.1
qdrant-client==1.15.0
scikit-learn==1.7.1
sentence-transformers==4.1.0
torch==2.7.1
uvicorn==0.35.0
gunicorn==23.0.0
python-multipart==0.0.20

